refactor(rendering): extract axis-agnostic leaf renderers into leaves.ts
Phase 1 of the column view plan. List view keeps its <table> and rowspan;
the four axis-agnostic leaf renderers move to a new shared module so the
column and time-grid containers can consume them without importing from
the list module.

src/rendering/leaves.ts (new, 596 lines) now owns:

  - renderDateWeather        date-column weather block
  - renderDateContent        weekday/day/month + colour precedence
  - renderLabel              event label pill
  - renderEventTitle         title + label + event weather
  - renderEventWeather       hourly-forecast event weather
  - renderEventContent       .event-content subtree (time/location/description)
  - renderTodayIndicator     today indicator (+ two private helpers)
  - EventContentParts        named interface for the six precomputed locals

render.ts drops 1120 -> 680 lines and keeps every wrapper: <td class="event">,
the accent/background/padding/position classes, and the rowspan geometry.
Those are exactly what differs between the list's table cell and a future
column card, so they stay with the container.

Contract: list-view DOM must be byte-identical. It is. The snapshot in
tests/__snapshots__/list-dom.test.ts.snap was never regenerated across all
four extractions, and the gate held at 86/86 after each one.

Two deviations from the plan as written, both internal and reversible:

  1. Leaves live in a new module rather than staying in render.ts. The plan
     requires column/time-grid to consume them "from their own containers";
     leaving them in place would make those containers import from the list.

  2. Targets 3 and 4 each moved as a cluster rather than the single named
     function. parseIndicatorPosition's only caller is renderTodayIndicator,
     and .event-content calls renderEventTitle, which calls renderLabel and
     renderEventWeather. Moving only the named function in either case would
     create a backwards leaves.ts -> render.ts import. renderEventTitle was
     already exported and had no consumers outside these two modules.

Also documents the whitespace trap found during the extraction: the gate's
serializer normalises whitespace between tags only, so indentation adjacent
to a text node is part of the oracle. Moved templates keep their original
absolute indentation verbatim, even where it looks wrong at the new nesting
depth. leaves.ts carries a header comment explaining why.

The no-output idiom guard in tests/list-dom.test.ts is repointed at leaves.ts
and gains a fifth assertion covering the date-column seam that remains in
render.ts - the Phase 2 cut point.

Copilot-Session: 549dfb35-2cb4-47ba-888c-44d766780ff8
