+++ src/rendering/render.ts
@@ -16,8 +16,7 @@ import * as Types from '../config/types';
 import * as Localize from '../translations/localize';
 import * as FormatUtils from '../utils/format';
 import * as EventUtils from '../utils/events';
-import * as Helpers from '../utils/helpers';
-import * as Weather from '../utils/weather';
+import * as Leaves from './leaves';
 
 //-----------------------------------------------------------------------------
 // MAIN CARD STRUCTURE RENDERING
@@ -299,41 +298,6 @@ function renderWeekRow(
   `;
 }
 
-//-----------------------------------------------------------------------------
-// EVENT CONTENT RENDERING HELPERS
-//-----------------------------------------------------------------------------
-
-/**
- * Render calendar label with support for text, emojis, images, and icons
- *
- * @param label - Label content from entity configuration
- * @param labelIconColor - Optional color for icon labels
- * @returns TemplateResult for the appropriate label type
- */
-function renderLabel(
-  label: string | undefined,
-  labelIconColor?: string,
-): TemplateResult | typeof nothing {
-  if (!label) return nothing;
-
-  // style attribute only if a color was provided
-  const styleAttr = labelIconColor ? `color: ${labelIconColor};` : nothing;
-
-  // Handle icons (mdi:, phu:, fas:, hass:, etc.)
-  if (Helpers.isIconValue(label)) {
-    return html`<ha-icon icon="${label}" class="label-icon" style=${styleAttr}></ha-icon>`;
-  }
-
-  // Handle image paths (either /local/ path or image file extension)
-  if (label.startsWith('/local/') || /\.(jpg|jpeg|png|gif|svg|webp)$/i.test(label)) {
-    return html`<img src="${label}" class="label-image"></img>`;
-  }
-
-  // Default: text/emoji (original behavior)
-  return html`<span class="calendar-label">${label}</span>`;
-}
-
-//-----------------------------------------------------------------------------
 // CONTENT GENERATION FUNCTIONS
 //-----------------------------------------------------------------------------
 
@@ -346,136 +310,6 @@ function renderLabel(
 function isWeekend(date: Date): boolean {
   const day = date.getDay();
   return day === 0 || day === 6; // 0 = Sunday, 6 = Saturday
-}
-
-/**
- * Parse position from CSS-like syntax (e.g., "10% 50%")
- * and convert to absolute positioning styles with centering transform
- *
- * @param position Position in CSS-like syntax ("x y" format)
- * @returns Style object with positioning properties
- */
-function parseIndicatorPosition(position: string): Record<string, string> {
-  // Default positioning styles
-  const positionStyles: Record<string, string> = {
-    position: 'absolute',
-    transform: 'translate(-50%, -50%)',
-  };
-
-  // Split the position string by whitespace
-  const parts = position.trim().split(/\s+/);
-
-  // Parse horizontal position (x)
-  if (parts.length >= 1) {
-    positionStyles.left = parts[0];
-  }
-
-  // Parse vertical position (y)
-  if (parts.length >= 2) {
-    positionStyles.top = parts[1];
-  } else {
-    // Default to vertically centered if only one value provided
-    positionStyles.top = '50%';
-  }
-
-  return positionStyles;
-}
-
-/**
- * Render the today indicator based on configuration
- *
- * @param config Calendar card configuration
- * @param isToday Whether the current day is today
- * @returns TemplateResult or nothing
- */
-function renderTodayIndicator(
-  config: Types.Config,
-  isToday: boolean,
-): TemplateResult | typeof nothing {
-  // Don't render anything if indicator is disabled or this isn't today
-  if (!config.today_indicator || !isToday) {
-    return nothing;
-  }
-
-  const indicatorValue = config.today_indicator;
-  const indicatorType = Helpers.getTodayIndicatorType(indicatorValue);
-
-  // If type is none, don't render anything
-  if (indicatorType === 'none') {
-    return nothing;
-  }
-
-  // Get position styles using CSS-like syntax
-  const positionStyles = parseIndicatorPosition(config.today_indicator_position);
-
-  // Render indicator based on type
-  return html`
-    <div class="today-indicator-container">
-      ${renderIndicatorByType(indicatorType, indicatorValue, positionStyles)}
-    </div>
-  `;
-}
-
-/**
- * Render specific indicator based on type
- */
-function renderIndicatorByType(
-  type: string,
-  value: string | boolean,
-  positionStyles: Record<string, string>,
-): TemplateResult | typeof nothing {
-  // Determine which icon to use based on type
-  let icon = '';
-
-  switch (type) {
-    case 'dot':
-      icon = 'mdi:circle';
-      break;
-    case 'pulse':
-      icon = 'mdi:circle';
-      break;
-    case 'glow':
-      icon = 'mdi:circle';
-      break;
-    case 'mdi':
-      // For custom MDI icons, use the value directly
-      icon = typeof value === 'string' ? value : 'mdi:circle';
-      break;
-    case 'image':
-      // For images, render an img tag instead
-      if (typeof value === 'string') {
-        return html`
-          <img 
-            src="${value}" 
-            class="today-indicator image"
-            style=${styleMap(positionStyles)}
-            alt="Today">
-          </img>`;
-      }
-      return nothing;
-    case 'emoji':
-      // For emojis, render a span with the emoji
-      if (typeof value === 'string') {
-        return html` <span class="today-indicator emoji" style=${styleMap(positionStyles)}>
-          ${value}
-        </span>`;
-      }
-      return nothing;
-    default:
-      return nothing;
-  }
-
-  // For all MDI-based indicators, render with the appropriate class
-  if (icon) {
-    return html` <ha-icon
-      icon="${icon}"
-      class="today-indicator ${type}"
-      style=${styleMap(positionStyles)}
-    >
-    </ha-icon>`;
-  }
-
-  return nothing;
 }
 
 /**
@@ -494,120 +328,10 @@ function renderDateColumn(
   isToday: boolean,
   weatherForecasts?: Types.WeatherForecasts,
 ): TemplateResult {
-  const isWeekendDay = date.getDay() === 0 || date.getDay() === 6;
+  // Weather is rendered by the shared leaf; the date block only positions it.
+  const weatherContent = Leaves.renderDateWeather(date, config, weatherForecasts);
 
-  // Start with base colors
-  let weekdayColor = config.weekday_color;
-  let dayColor = config.day_color;
-  let monthColor = config.month_color;
-
-  // Apply weekend styling if applicable and defined
-  if (isWeekendDay) {
-    weekdayColor = config.weekend_weekday_color || weekdayColor;
-    dayColor = config.weekend_day_color || dayColor;
-    monthColor = config.weekend_month_color || monthColor;
-  }
-
-  // Apply today styling if applicable and defined (takes precedence)
-  if (isToday) {
-    weekdayColor = config.today_weekday_color || weekdayColor;
-    dayColor = config.today_day_color || dayColor;
-    monthColor = config.today_month_color || monthColor;
-  }
-
-  // Get translations for the current language
-  const translations = Localize.getTranslations(language);
-
-  // Get formatted date parts from translations
-  const weekday = translations.daysOfWeek[date.getDay()];
-  const day = date.getDate();
-  const month = translations.months[date.getMonth()];
-
-  // Add weather if configured
-  const showDateWeather =
-    (config.weather?.position === 'date' || config.weather?.position === 'both') &&
-    config.weather?.entity;
-
-  let weatherContent: TemplateResult | typeof nothing = nothing;
-
-  if (showDateWeather && weatherForecasts?.daily) {
-    const dailyForecast = Weather.findDailyForecast(date, weatherForecasts.daily);
-
-    if (dailyForecast) {
-      // Get options from date-specific config
-      const dateConfig = config.weather?.date || {};
-      const showConditions = dateConfig.show_conditions !== false;
-      const showHighTemp = dateConfig.show_high_temp !== false;
-      const showUvIndex =
-        dateConfig.show_uv_index === true &&
-        dailyForecast.uv_index !== undefined &&
-        dailyForecast.uv_index >= (dateConfig.uv_index_threshold ?? 0);
-      const showLowTemp =
-        dateConfig.show_low_temp === true && !showUvIndex && dailyForecast.templow !== undefined;
-
-      // Get styling from config
-      const iconSize = dateConfig.icon_size || '14px';
-      const fontSize = dateConfig.font_size || '12px';
-      const color = dateConfig.color || 'var(--primary-text-color)';
-
-      weatherContent = html`
-        <div class="weather" style="font-size: ${fontSize}; color: ${color};">
-          ${showConditions
-            ? html`
-                <ha-icon
-                  .icon=${dailyForecast.icon}
-                  style="--mdc-icon-size: ${iconSize};"
-                ></ha-icon>
-              `
-            : nothing}
-          ${showHighTemp
-            ? html` <span class="weather-temp-high">${dailyForecast.temperature}°</span> `
-            : nothing}
-          ${showLowTemp
-            ? html` <span class="weather-temp-low">/${dailyForecast.templow}°</span> `
-            : nothing}
-          ${showUvIndex
-            ? html` <span class="weather-uv-index">UV${dailyForecast.uv_index}</span> `
-            : nothing}
-        </div>
-      `;
-    }
-  }
-
-  return html`
-    <div
-      class="weekday"
-      style=${styleMap({
-        'font-size': config.weekday_font_size,
-        color: weekdayColor,
-      })}
-    >
-      ${weekday}
-    </div>
-    <div
-      class="day"
-      style=${styleMap({
-        'font-size': config.day_font_size,
-        color: dayColor,
-      })}
-    >
-      ${day}
-    </div>
-    ${config.show_month
-      ? html`
-          <div
-            class="month"
-            style=${styleMap({
-              'font-size': config.month_font_size,
-              color: monthColor,
-            })}
-          >
-            ${month}
-          </div>
-        `
-      : nothing}
-    ${weatherContent}
-  `;
+  return Leaves.renderDateContent(date, config, language, isToday, weatherContent);
 }
 
 /**
@@ -921,6 +645,16 @@ export function renderEvent(
     'past-event': isPastEvent,
   };
 
+  // Everything the event body needs that it must not recompute for itself.
+  const contentParts: Leaves.EventContentParts = {
+    eventTime,
+    eventLocation,
+    eventDescription,
+    shouldShowTime,
+    countdownStr,
+    progressPercentage,
+  };
+
   return html`
     <tr>
       ${index === 0
@@ -931,7 +665,7 @@ export function renderEvent(
               style="position: relative;"
             >
               ${renderDateColumn(dayDate, config, language, isToday, weatherForecasts)}
-              ${renderTodayIndicator(config, isToday)}
+              ${Leaves.renderTodayIndicator(config, isToday)}
             </td>
           `
         : ''}
@@ -939,182 +673,8 @@ export function renderEvent(
         class=${classMap(eventClasses)}
         style="border-inline-start: var(--calendar-card-line-width-vertical) solid ${entityAccentColor}; background-color: ${entityAccentBackgroundColor};"
       >
-        <div class="event-content">
-          ${renderEventTitle(event, config, weatherForecasts)}
-          <div class="time-location">
-            ${shouldShowTime
-              ? html`
-                  <div class="time">
-                    <div class="time-actual">
-                      <ha-icon icon="mdi:clock-outline"></ha-icon>
-                      <span>${eventTime}</span>
-                    </div>
-                    ${countdownStr
-                      ? html`<div class="time-countdown">${countdownStr}</div>`
-                      : progressPercentage !== null && config.show_progress_bar
-                        ? html`
-                            <div class="progress-bar">
-                              <div
-                                class="progress-bar-filled"
-                                style="width: ${progressPercentage}%"
-                              ></div>
-                            </div>
-                          `
-                        : nothing}
-                  </div>
-                `
-              : countdownStr
-                ? html`
-                    <div class="time">
-                      <div class="time-actual"></div>
-                      <div class="time-countdown">${countdownStr}</div>
-                    </div>
-                  `
-                : progressPercentage !== null && config.show_progress_bar
-                  ? html`
-                      <div class="time">
-                        <div class="time-actual"></div>
-                        <div class="progress-bar">
-                          <div
-                            class="progress-bar-filled"
-                            style="width: ${progressPercentage}%"
-                          ></div>
-                        </div>
-                      </div>
-                    `
-                  : nothing}
-            ${eventLocation
-              ? html`
-                  <div class="location">
-                    <ha-icon icon="mdi:map-marker-outline"></ha-icon>
-                    <span>${eventLocation}</span>
-                  </div>
-                `
-              : ''}
-            ${eventDescription
-              ? html`
-                  <div class="description">
-                    <ha-icon icon="mdi:information-outline"></ha-icon>
-                    <span>${eventDescription}</span>
-                  </div>
-                `
-              : ''}
-          </div>
-        </div>
+        ${Leaves.renderEventContent(event, config, contentParts, weatherForecasts)}
       </td>
     </tr>
-  `;
-}
-
-/**
- * Render an event title with optional label and weather data
- */
-export function renderEventTitle(
-  event: Types.CalendarEventData,
-  config: Types.Config,
-  weatherForecasts?: Types.WeatherForecasts,
-): TemplateResult {
-  const isEmptyDay = !!event._isEmptyDay;
-  // The checkmark reads as "nothing on" and only suits the default text.
-  // A user-supplied string ("Leftovers") already carries its own meaning,
-  // so the prefix is dropped whenever one is configured.
-  const showEmptyDayCheckmark = isEmptyDay && !event._isCustomEmptyText;
-  const entityColor = isEmptyDay
-    ? 'var(--calendar-card-empty-day-color)'
-    : event._matchedConfig?.color || config.event_color;
-
-  const entityLabel = EventUtils.getEntityLabel(event._entityId, config, event);
-
-  // label_icon_color from the matched entity config
-  const labelIconColor = event._matchedConfig?.label_icon_color;
-
-  return html`
-    <div class="summary-row">
-      <div class="summary">
-        ${entityLabel ? renderLabel(entityLabel, labelIconColor) : nothing}
-        <span
-          class="event-title ${isEmptyDay ? 'empty-day-title' : ''}"
-          style="color: ${entityColor}"
-        >
-          ${showEmptyDayCheckmark ? `✓ ${event.summary}` : event.summary}
-        </span>
-      </div>
-      ${renderEventWeather(event, config, weatherForecasts)}
-    </div>
-  `;
-}
-
-/**
- * Render weather information for an event
- */
-function renderEventWeather(
-  event: Types.CalendarEventData,
-  config: Types.Config,
-  weatherForecasts?: Types.WeatherForecasts,
-): TemplateResult {
-  // Only render if weather is enabled for events
-  const showEventWeather =
-    config.weather?.entity &&
-    (config.weather.position === 'event' || config.weather.position === 'both');
-
-  if (!showEventWeather || !weatherForecasts?.hourly) {
-    return html``;
-  }
-
-  // Check if this is a timed event (has dateTime) that has ended
-  if (event.end?.dateTime) {
-    const now = new Date();
-    const eventEndTime = new Date(event.end.dateTime);
-
-    // If event has ended, don't show weather
-    if (eventEndTime < now) {
-      return html``;
-    }
-  }
-
-  // Get options from event-specific config
-  const eventConfig = config.weather?.event || {};
-
-  // Find the appropriate forecast - pass both hourly and daily forecasts
-  const forecast = Weather.findForecastForEvent(
-    event,
-    weatherForecasts.hourly,
-    weatherForecasts.daily,
-    eventConfig.daily_forecast_fallback !== false,
-  );
-
-  if (!forecast) {
-    return html``;
-  }
-
-  const showConditions = eventConfig.show_conditions !== false;
-  const showTemp = eventConfig.show_temp !== false;
-  const showUvIndex =
-    eventConfig.show_uv_index === true &&
-    forecast.uv_index !== undefined &&
-    forecast.uv_index >= (eventConfig.uv_index_threshold ?? 0);
-
-  // Get styling from config
-  const iconSize = eventConfig.icon_size || '14px';
-  const fontSize = eventConfig.font_size || '12px';
-  const color = eventConfig.color || 'var(--secondary-text-color)';
-
-  // Render weather with position-specific options
-  return html`
-    <div class="event-weather">
-      ${showConditions
-        ? html`<ha-icon .icon=${forecast.icon} style="--mdc-icon-size: ${iconSize};"></ha-icon>`
-        : nothing}
-      ${showTemp
-        ? html`<span style="font-size: ${fontSize}; color: ${color};">
-            ${forecast.temperature}°
-          </span>`
-        : nothing}
-      ${showUvIndex
-        ? html`<span class="weather-uv-index" style="font-size: ${fontSize}; color: ${color};">
-            UV${forecast.uv_index}
-          </span>`
-        : nothing}
-    </div>
   `;
 }
