+++ tests/list-dom.test.ts
@@ -605,29 +605,37 @@ describe('list view DOM', () => {
 
   // Source-text guard, not a DOM assertion — deliberately.
   //
-  // render.ts uses three interchangeable "render nothing" idioms: `''`, `nothing`,
+  // The renderers use three interchangeable "render nothing" idioms: `''`, `nothing`,
   // and an empty html`` template. The rendered DOM cannot tell `''` from `nothing`,
   // so no behavioural test can pin them; only reading the source can.
   //
-  // This test IS EXPECTED TO FAIL during the Phase 1 extraction, because the code it
-  // matches will move out of render.ts into leaf renderers. That failure is the point:
-  // it forces a conscious re-read of each idiom at the moment it is most likely to be
-  // silently normalised. When it fails, repoint the regexes at the new file and confirm
-  // each idiom survived unchanged. Do not delete it, and do not "fix" it by relaxing a
-  // regex to match whatever the new code happens to say.
+  // Phase 1 moved the event leaves out of render.ts into leaves.ts. This guard failed
+  // at that moment by design, and each idiom below was re-read and confirmed unchanged
+  // before the regexes were repointed. It is expected to fail again at each later
+  // extraction seam, for the same reason. When it does: repoint, having first confirmed
+  // every idiom survived byte-for-byte. Do not delete it, and do not "fix" it by relaxing
+  // a regex to match whatever the new code happens to say.
+  //
+  // Phase 2 will cut the `${index === 0 ? html`…` : ''}` seam in render.ts, when the
+  // column container grows its own date rendering. That idiom is asserted here too so
+  // the same forcing function applies to it.
   it('preserves no-output idioms at extraction seams', () => {
     const renderSource = readFileSync(`${process.cwd()}/src/rendering/render.ts`, 'utf8');
-    const eventWeatherSource = renderSource.slice(
-      renderSource.indexOf('function renderEventWeather'),
+    const leavesSource = readFileSync(`${process.cwd()}/src/rendering/leaves.ts`, 'utf8');
+    const eventWeatherSource = leavesSource.slice(
+      leavesSource.indexOf('function renderEventWeather'),
     );
 
-    expect(renderSource).toMatch(
+    expect(leavesSource).toMatch(
       /\$\{eventLocation\s*\? html`[\s\S]*?`\s*: ''\}\s*\$\{eventDescription/,
     );
-    expect(renderSource).toMatch(/\$\{eventDescription\s*\? html`[\s\S]*?`\s*: ''\}\s*<\/div>/);
-    expect(renderSource).toMatch(
+    expect(leavesSource).toMatch(/\$\{eventDescription\s*\? html`[\s\S]*?`\s*: ''\}\s*<\/div>/);
+    expect(leavesSource).toMatch(
       /progressPercentage !== null && config\.show_progress_bar[\s\S]*?: nothing\}\s*\$\{eventLocation/,
     );
     expect(eventWeatherSource.match(/return html``;/g)).toHaveLength(3);
+
+    // The list row's date cell renders `''`, not `nothing`, on every row but the first.
+    expect(renderSource).toMatch(/\$\{index === 0\s*\? html`[\s\S]*?`\s*: ''\}/);
   });
 });
